Got rid of new GCC 4.3.2 warnings.
git-svn-id: svn+ssh://svn.code.sf.net/p/x3270/x3270/trunk@433 12e4f418-770b-42d6-82dd-bd6eefd7344d

+++ x3270/tcl3270/tcl3270.c
@@ -630,7 +630,7 @@ x3270_cmd(ClientData clientData, Tcl_Interp *interp, int objc,
 			trace_event(" -> \"%s\"",
 			    scatv(s_trunc, buf, sizeof(buf)));
 			if (strlen(s) > TRUNC_LEN)
-				trace_event("...(%d chars)", strlen(s));
+				trace_event("...(%d chars)", (int)strlen(s));
 		}
 		trace_event("\n");
 	}
